requirements.txt
numpy==1.17.3
opencv-python==4.4.0.46
pandas==1.1.5
pascal-voc-writer==0.1.4
pillow==8.0.1
protobuf==3.13.0.1
python==3.7.9
tensorboard==2.4.0
tensorflow==2.3.1
shapely==1.7.1
